Speed up refresh
The data line is now changed when the clock line falls. That saves one
cycle but means that the clock line must be high when CLOCK_DATA() is
called.

The BLANK_ signal is now LOW all the time.

The LED pattern is a little more interesting. Red brightens, Green
fades, blue brightens.

The dimness pattern is updated to spread across the entire cycle instead
of being bursty up front. Using a lookup table.